requirements.txt
urllib
bs4
PySocks
chardet